+++ test/integration/src/components/common/PhotoCaption/PhotoCaptionThumbnail.js
@@ -3,7 +3,7 @@ import PhotoCaption from './PhotoCaption.js';
 
 import url from './mountains.jpg';
 
-export default class PhotoCaptionThumbView extends React.Component {
+export default class PhotoCaptionThumbnail extends React.Component {
   render() {
     return <div>
       <PhotoCaption url={url} caption={"Photo by Brandon Lam"} width={300} />